// src/types/configuration.ts
export enum ConfigRoutePath {
  Colorimetric = '/config/colorimetric',
  Assortment = '/config/assortment',
  Colorants = '/config/colorants',
  Users = '/config/users',
  Groups = '/config/groups',
  AccessControl = '/config/access-control',
  BridgeApp = '/config/bridge',
  About = '/config/about',
  Formulation = '/config/formulation',
}